requirements.txt
# Windows 7 Compatible Requirements
# Note: Use Python 3.8 for Windows 7 compatibility
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
jdatetime==4.1.1
openpyxl==3.1.2
pandas==1.5.3
numpy==1.24.3
et-xmlfile==1.1.0
python-dateutil==2.8.2
jalali-core==1.0.0
cachetools==5.3.0
protobuf==3.20.3
ippanel==2.0.7
python-dotenv==1.0.0
pyinstaller==5.7.0
auto-py-to-exe==2.20.0
